feat: Added PhysiologicalMotor for citizens

+++ src/main/java/com/caosmos/citizens/application/dto/CitizenSummaryDto.java
@@ -10,7 +10,7 @@ public record CitizenSummaryDto(
     double z,
     String state,
     String currentGoal,
-    int vitality
+    double vitality
 ) {
 
 }
