+++ src/main/java/com/caosmos/common/application/telemetry/BiometricsEntry.java
@@ -5,8 +5,8 @@ import java.util.UUID;
 public record BiometricsEntry(
     UUID entityId,
     long tick,
-    int vitality,
-    int energy
+    double vitality,
+    double energy
 ) {
 
 }
